# src/backend/rest/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: FridoLean.io
host: localhost:3001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  '/api/v1/projects/{projectID}/timeline':
    x-swagger-router-controller: timeline
    get:
      tags:
        - timeline
        - project
      description: Returns a list of the tagged states stored in the timeline
      operationId: getTimeline
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          description: ID of a project
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            properties:
              states:
                type: array
                items:
                  $ref: '#/definitions/TimelineState'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      tags:
        - timeline
        - project
      description: Store the current project state in the timeline
      operationId: postTimeline
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          description: ID of a project
          required: true
          type: string
          format: uuid
        - name: tag
          in: body
          description: Specifies the tag which is stored together with the project state
          required: true
          schema:
            type: object
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/TimelineState'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}/timeline/{stateID}':
    x-swagger-router-controller: timeline
    get:
      tags:
        - timeline
        - project
      description: Get the content of a specific tagged timeline state
      operationId: getTimelineState
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          description: ID of a project
          required: true
          type: string
          format: uuid
        - name: stateID
          in: path
          description: ID of a specific timeline state
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            properties:
              stateInformation:
                $ref: '#/definitions/TimelineState'
              content:
                $ref: '#/definitions/ProjectContent'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}/timeline/{stateID}/restore':
    x-swagger-router-controller: timeline
    put:
      tags:
        - timeline
        - project
      description: >-
        Restore a project state saved in the timeline (copy it to the current
        project state)
      operationId: restoreState
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          description: ID of a project
          required: true
          type: string
          format: uuid
        - name: stateID
          in: path
          description: ID of a specific timeline state
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            properties:
              stateInformation:
                $ref: '#/definitions/TimelineState'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project or State not found
          schema:
            $ref: '#/definitions/NotFoundErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}/timeline/{stateID}/copy':
    x-swagger-router-controller: timeline
    post:
      tags:
        - timeline
        - project
      description: >-
        Copy a project state (including all previous timeline states) and create
        a new project (for the user).
      operationId: copyState
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          description: ID of a project
          required: true
          type: string
          format: uuid
        - name: stateID
          in: path
          description: ID of a specific timeline state
          required: true
          type: string
          format: uuid
        - name: data
          required: true
          description: Metadata of the new project
          in: body
          schema:
            $ref: '#/definitions/CreateProject'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ProjectMeta'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project or State not found
          schema:
            $ref: '#/definitions/NotFoundErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  CreateProject:
    properties:
      title:
        type: string
        default: Untitled
        description: Title of the project
      description:
        description: Description of the project
        type: string
      visibility:
        type: string
        description: PUBLIC || PRIVATE
...
  ProjectMeta:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      visibility:
        type: string
        description: PUBLIC || PRIVATE
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      description:
        type: string
      members:
        type: array
        items:
          properties:
            userID:
              type: string
            email:
              type: string
            rights:
              type: string
  ProjectContent:
    properties:
      canvases:
        type: array
        items:
          $ref: '#/definitions/Canvas'
  TimelineState:
    properties:
      id:
        type: string
        format: uuid
      tag:
        type: string
      timestamp:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
...
  Entry:
    properties:
      _id:
        type: string
      type:
        type: string
      content:
        $ref: '#/definitions/Content'
  Content:
    properties:
      title:
        type: string
      text:
        type: string
      reference:
        type: string
      target:
        type: string
  BuildingBlock:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
        description: |
          BuisinessModelCanvas: [
            KeyPartners,
            KeyActivities,
            KeyResources,
            CustomerRelation,
            Channels,
            CustomerSegments,
            CostStructure,
            RevenueStreams,
            ValueProposition
          ],                                    

          ValuePropositionCanvas: [      
            ProductsAndServices,
            GainCreators,
            PainRelievers,
            Gains,
            Pains ,                                
            CustomerJobs
          ],                            

          LeanCanvas: [
            Problem,                                  
            Solution,                                   
            KeyMetrics,                                      
            UniqueValueProposition,                                                   
            UnfairAdvantages,                                           
            Channels,                                   
            CustomerSegments,                                           
            CostStructure,                                         
            RevenueStreams
          ],                                 

          CustomerJourney: [
            Advertisement,                                        
            PreSocialMedia,                                              
            PreWordOfMouth,                                               
            Expectations,                                       
            PastExperiences,                                           
            ServiceJourney,                                          
            Experiences,                                      
            RelationshipMenage,                                                  
            PostSocialMedia,                                               
            PostWordOfMouth,                                                
            DisSatisfaction
          ]
      buildingBlockType:
        type: string
      layoutEntry:
        properties:
          _id:
            type: string
            format: uuid
          x:
            type: number
          'y':
            type: number
          width:
            type: number
          height:
            type: number
      entries:
        type: array
        items:
          $ref: '#/definitions/Entry'
...
  Canvas:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      canvasType:
        type: string
        description: BUSINESS_MODEL || LEAN || VALUE_PROPOSITION || CUSTOMER_JOURNEY
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      buildingBlocks:
        type: array
        items:
          $ref: '#/definitions/BuildingBlock'
...
  NotFoundErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        description: Describes which instance could not be found
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...